move entrance & grad year calculation to utilities.ts
